deps: update all examples to use bdk_wallet 1.0.0
