+++ examples/rust/descriptors/Cargo.toml
@@ -4,4 +4,4 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-bdk_wallet = { version = "=1.0.0-beta.5", features = ["keys-bip39"] }
+bdk_wallet = { version = "1.0.0", features = ["keys-bip39"] }
