+++ examples/rust/full-wallet/Cargo.toml
@@ -4,6 +4,6 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-bdk_wallet = { version = "=1.0.0-beta.5", features = ["keys-bip39", "rusqlite"] }
-bdk_esplora = { version = "=0.19.0", features = ["blocking"] }
+bdk_wallet = { version = "1.0.0", features = ["keys-bip39", "rusqlite"] }
+bdk_esplora = { version = "0.20.1", features = ["blocking"] }
 anyhow = "1"
