+++ examples/rust/memory-wallet/.gitignore
@@ -1,0 +1,1 @@
+target/
