+++ examples/rust/migrate-version/Cargo.toml
@@ -7,5 +7,5 @@ edition = "2021"
 [dependencies]
 anyhow = "1"
 bdk = { version = "0.30", features = ["sqlite"] }
-bdk_wallet = { version = "=1.0.0-beta.5", features = ["rusqlite"] }
+bdk_wallet = { version = "1.0.0", features = ["rusqlite"] }
 # --8<-- [end:deps]
