+++ examples/rust/quickstart/Cargo.toml
@@ -4,5 +4,5 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-bdk_wallet = { version = "=1.0.0-beta.5" }
-bdk_esplora = { version = "=0.19.0", features = ["blocking"] }
+bdk_wallet = { version = "1.0.0" }
+bdk_esplora = { version = "0.20.1", features = ["blocking"] }
