+++ examples/rust/syncing/electrum/.gitignore
@@ -1,0 +1,2 @@
+target/
+electrumexample.db
