+++ examples/rust/syncing/electrum/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "electrum"
+version = "0.1.0"
+edition = "2021"
+
+# --8<-- [start:deps]
+[dependencies]
+bdk_wallet = { version = "1.0.0" }
+bdk_electrum = { version = "0.20.1" }
+# --8<-- [end:deps]
