+++ examples/rust/syncing/esplora/.gitignore
@@ -1,0 +1,2 @@
+target/
+esploraexample.db
