+++ examples/rust/syncing/kyoto/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "kyoto"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+bdk_kyoto = { version = "0.6.0", default-features = false, features = ["rusqlite", "wallet", "callbacks"] }
+bdk_wallet = { version = "1.0.0" }
+tokio = { version = "1.37", features = ["full"], default-features = false }
