+++ examples/rust/syncing/rpc/.gitignore
@@ -1,0 +1,2 @@
+target/
+rpcexample.db
