+++ examples/rust/syncing/rpc/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "rpc"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+bdk_wallet = { version = "1.0.0" }
+bdk_bitcoind_rpc = {  version = "0.17.1" }
