--- examples/syncing/electrum/.gitignore
@@ -1,2 +1,0 @@
-target/
-electrumexample.db
