--- examples/syncing/esplora/.gitignore
@@ -1,2 +1,0 @@
-target/
-esploraexample.db
