--- examples/syncing/esplora/Cargo.toml
@@ -1,8 +1,0 @@
-[package]
-name = "esplora"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-bdk_wallet = { version = "=1.0.0-beta.1" }
-bdk_esplora = { version = "=0.16.0", features = ["blocking"] }
