--- examples/syncing/rpc/.gitignore
@@ -1,2 +1,0 @@
-target/
-rpcexample.db
