--- examples/syncing/rpc/Cargo.toml
@@ -1,8 +1,0 @@
-[package]
-name = "rpc"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-bdk_wallet = { version = "=1.0.0-beta.1" }
-bdk_bitcoind_rpc = {  version = "0.13.0" }
